requirement.txt
# Messaging
kafka-python==2.0.2
